+++ Strags/PK10/ProbWaveSelectStragClass.cs
@@ -71,7 +71,7 @@ namespace WolfInv.com.Strags
         /// <param name="RestCash"></param>
         /// <param name="cc"></param>
         /// <returns></returns>
-        public override Int64 getChipAmount(double RestCash, ChanceClass cc, AmoutSerials ams)
+        public override double getChipAmount(double RestCash, ChanceClass cc, AmoutSerials ams)
         {
             ////if (this.UseAmountList.ContainsKey(this.LastUseData.LastData.Expect))
             ////{
