+++ Strags/PK10/strag_CommJumpClass.cs
@@ -38,7 +38,7 @@ namespace WolfInv.com.Strags
             return true;
         }
 
-        public override long getChipAmount(double RestCash, ChanceClass cc, AmoutSerials amts)
+        public override double getChipAmount(double RestCash, ChanceClass cc, AmoutSerials amts)
         {
             return getDefaultChipAmount(RestCash, cc, amts);
             return amts.Serials[0][HoldCnt-1];
