+++ Strags/PK10/strag_CommLongMissBackBalance.cs
@@ -120,7 +120,7 @@ namespace WolfInv.com.Strags
             return scs;
         }
 
-        public override Int64 getChipAmount(double RestCash, ChanceClass cc, AmoutSerials ams)
+        public override double getChipAmount(double RestCash, ChanceClass cc, AmoutSerials ams)
         {
             if(cc.IncrementType==  InterestType.SimpleInterest)
                 return (int)Math.Floor(this.CommSetting.InitCash*0.01);
