+++ Strags/PK10/strag_PoissonRandomClass.cs
@@ -44,7 +44,7 @@ namespace WolfInv.com.Strags
             return true;
         }
 
-        public override long getChipAmount(double RestCash, ChanceClass cc, AmoutSerials amts)
+        public override double getChipAmount(double RestCash, ChanceClass cc, AmoutSerials amts)
         {
             return (Int64)Math.Floor(RestCash*0.001);//ProbMath.GetFactorial
         }
